requirements.txt
torch
torchvision
streamlit
opencv-python
Pillow
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
pytesseract
matplotlib
